機能毎にクラス分け。
