+++ CalcLib/Moriguchi/CalcSvcMoriguchi.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace CalcLib.Moriguchi
 {
-    internal class CalcSvcMoriguchi : ICalcSvcEx
+    public class CalcSvcMoriguchi : ICalcSvcEx
     {
         class OpeNameHelper
         {
@@ -22,7 +22,7 @@ namespace CalcLib.Moriguchi
             public static string Get(CalcButton? opeButton) => opeButton.HasValue ? OpeTextTable[opeButton.Value] : "";
         }
 
-        class CalcContextMoriguchi : ICalcContext
+        public class CalcContextMoriguchi : ICalcContext
         {
             /// <summary>
             /// 左辺の値
@@ -57,7 +57,20 @@ namespace CalcLib.Moriguchi
             /// おみくじ
             /// </summary>
             public string[] omikuji = { "大吉", "中吉", "小吉", "凶　" };
+
+            public void Clear()
+            {
+                Buffer = "0";
+                Value = null;
+                Operation = null;
+            }
+
         }
+
+        /// <summary>
+        /// サービスの入れ物
+        /// </summary>
+        static ISubSvc svc;
 
         public virtual ICalcContext CreateContext() => new CalcContextMoriguchi();
 
@@ -84,238 +97,29 @@ namespace CalcLib.Moriguchi
             var ctx = ctx0 as CalcContextMoriguchi;
             Debug.WriteLine($"Button Clicked {btn}, context={ctx}");
 
-            if (ctx.Mode)
-            {    //おみくじモード時
-                OmikujiMethod(btn, ctx);
-            }
-            else
+            //defaultでは電卓モード
+            if (svc == null)
             {
-                //電卓モード時
-                CalcMethod(btn, ctx);
-            }
-        }
-
-        /// <summary>
-        /// おみくじモード時の動作
-        /// </summary>
-        /// <param name="btn"></param>
-        /// <param name="ctx"></param>
-        private void OmikujiMethod(CalcButton btn, CalcContextMoriguchi ctx)
-        {
-            //押下ボタン判定
-            switch (btn)
-            {
-                case CalcButton.Btn1:
-                case CalcButton.Btn2:
-                case CalcButton.Btn3:
-                case CalcButton.Btn4:
-                    OpenOmikuji(btn, ctx);
-                    //おみくじを1回でも引いたら電卓モードへ
-                    ctx.Mode = false;
-                    break;
-
-                //電卓モードへ戻る時
-                case CalcButton.BtnClear:
-                case CalcButton.BtnClearEnd:
-                case CalcButton.BtnExt2:
-                    ctx.Mode = false;
-                    ctx.Buffer = "0";
-                    ctx.Value = null;
-                    ctx.Operation = null;
-                    break;
-                //関係ないボタン押下時:何もしない
-            }
-        }
-
-        /// <summary>
-        /// おみくじの開示
-        /// </summary>
-        /// <param name="btn"></param>
-        /// <param name="ctx"></param>
-        private void OpenOmikuji(CalcButton btn, CalcContextMoriguchi ctx)
-        {
-            //おみくじ配列のシャッフル
-            var test = ctx.omikuji.OrderBy(x => Guid.NewGuid()).ToArray();
-
-            //開示表示
-            ctx.Value = $"本日の運勢は「{test[(int)btn - 1]}」です";
-            ctx.Buffer = null;
-            foreach (var kekka in test.Select(x => x))
-            {
-                ctx.Buffer += kekka + " ";
-            };
-            Chomp(ctx);
-        }
-
-        private static void Chomp(CalcContextMoriguchi ctx)
-           => ctx.Buffer = ctx.Buffer.Remove(ctx.Buffer.Length - 1);
-
-        private void CalcMethod(CalcButton btn, CalcContextMoriguchi ctx)
-        {
-            if (!string.IsNullOrEmpty(ctx.Value) && ctx.Value.StartsWith("本"))
-            {
-                ctx.Value = null;
-                ctx.Buffer = null;
+                svc = new CalcClass();
+                svc.Init(ctx);
             }
 
             switch (btn)
             {
-                //演算子
-                case CalcButton.BtnPlus:
-                case CalcButton.BtnMinus:
-                case CalcButton.BtnDivide:
-                case CalcButton.BtnMultiple:
-                    //小数点押下直後に演算子を押下すると小数点を削除する
-                    if (ctx.Buffer?.EndsWith(".") == true)
-                    {
-                        char[] dot = { '.' };
-                        ctx.Buffer = ctx.Buffer.TrimEnd(dot);
-                    }
-                    OnOpeButtonClick(ctx, btn);   // 演算子ボタン押下時の処理
-                    break;
-
-                //クリア
-                case CalcButton.BtnClear:
-                    ctx.Buffer = null;
-                    ctx.Value = null;
-                    ctx.Operation = null;
-                    //ctx.SubDisplayText = null;
-                    break;
-                case CalcButton.BtnClearEnd:
-                    break;
-                //バックスペース
-                case CalcButton.BtnBS:
-                    if (!string.IsNullOrEmpty(ctx.Buffer))
-                    {
-                        Chomp(ctx);
-                    }
-                    break;
-
-                //パーセント(BtnExt1)押下時
-                case CalcButton.BtnExt1:
-                    //Valueがnullでない時、valueのbuffer%をbufferに入れる
-                    //TODO:Valueに値が入っていると%ボタンを押下する度に計算してしまう
-                    if (!string.IsNullOrEmpty(ctx.Value))
-                    {
-                        var val = double.Parse(ctx.Value);
-                        var buf = double.Parse(ctx.Buffer);
-                        ctx.Buffer = (val * (buf / 100)).ToString();
-                    }
-                    break;
-
-                //おみくじモードボタン押下時
                 case CalcButton.BtnExt2:
-                    ctx.Mode = true;
-                    ctx.Value = "おみくじを選択して下さい";
-                    ctx.Operation = btn;
-                    ctx.Buffer = "[1 ] [2 ] [3 ] [4 ]";
+                    svc = new OmikujiClass();
+                    svc.Init(ctx);
                     break;
-
-                //小数点押下時
-                case CalcButton.BtnDot:
-                    if (!ctx.Buffer?.EndsWith(".") == true) ctx.Buffer += ".";
-                    break;
-
-                //計算
-                case CalcButton.BtnEqual:
-                    if (!string.IsNullOrEmpty(ctx.Buffer) && !string.IsNullOrEmpty(ctx.Value) && ctx.Operation != null)
-                    {
-                        ExecCalcuration(ctx, ctx.Operation.Value);
-                        ctx.Operation = null;
-                        //ctx.SubDisplayText = null;
-                    }
-                    break;
-
-                //入力数値取得
                 default:
-                    if (ctx.Reset)
-                    {
-                        ctx.Buffer = null;
-                        ctx.Reset = false;
-                    }
-                    ctx.Buffer += (int)btn;
                     break;
             }
 
-        }
+            var ret = svc.OnClick(ctx, btn);
 
-        /// <summary>
-        /// 演算子ボタン押下時の処理
-        /// </summary>
-        /// <param name="ctx"></param>
-        private void OnOpeButtonClick(CalcContextMoriguchi ctx, CalcButton btn)
-        {
-            ctx.Operation = btn;
-            //ctx.SubDisplayText += ctx.Buffer;
-
-            if (string.IsNullOrEmpty(ctx.Value))
+            if (!ret)
             {
-                // 左辺が未入力の時、Bufferの値を左辺とする
-                ctx.Value = ctx.Buffer;
-                ctx.Buffer = null;
-                //ctx.SubDisplayText += ctx.Buffer + OpeNameHelper.Get(btn);
+                svc = null;
             }
-            else if (!string.IsNullOrEmpty(ctx.Buffer))
-            {
-                // 左辺が入力済みで、Bufferが入力済みの時、計算処理を実行する
-                var x = ExecCalcuration(ctx, btn);
-
-                // 続けて計算できるよう実行結果を左辺にセットする。
-                ctx.Value = x;
-                ctx.Buffer = null;
-                //ctx.SubDisplayText += OpeNameHelper.Get(btn);
-            }
-        }
-
-        /// <summary>
-        /// 計算処理を実行します
-        /// </summary>
-        /// <param name="ctx"></param>
-        /// <returns></returns>
-        private string ExecCalcuration(CalcContextMoriguchi ctx, CalcButton operation)
-        {
-            string x;
-            {
-                x = string.Format("{0:#,0.#############}",Calc(ctx.Value, ctx.Buffer, operation));
-                ctx.Reset = true;
-                ctx.Buffer = x;
-                ctx.Value = null;
-            }
-
-            return x;
-        }
-
-        /// <summary>
-        /// 計算処理
-        /// </summary>
-        /// <param name="Value1"></param>
-        /// <param name="Value2"></param>
-        /// <param name="Ope"></param>
-        /// <returns></returns>
-        public double Calc(string Value, string Buffer, CalcButton Ope)
-        {
-            double answer = 0;
-
-            var val1 = double.Parse(Value);
-            var val2 = double.Parse(Buffer);
-
-            if (Ope == CalcButton.BtnPlus)
-            {
-                answer = val1 + val2;
-            }
-            else if (Ope == CalcButton.BtnMinus)
-            {
-                answer = val1 - val2;
-            }
-            else if (Ope == CalcButton.BtnMultiple)
-            {
-                answer = val1 * val2;
-            }
-            else if (Ope == CalcButton.BtnDivide)
-            {
-                answer = val1 / val2;
-            }
-            return answer;
         }
     }
 }
